refactor: renamed player.js to playerInfor.js
