+++ src/phaser/player/playerInfor.js
